Change the way that qcd.py is using optionparser.
I like the idea behind optionparser.py, but it seems a bit rigid.

So, I have changed the impl so that from the main program (qcd) a global dict
is used to contain all the options for later use. In this way, is a bit more
explicit that a function is accessing to a global stash rather than making
magic. Now the code says:

    options = dict()
    def initialize_database (writeable = False):
       file = expanduser ( options["dbfile"].value )

Instead of:

    def initialize_database (writeable = False):
       file = expanduser (file_option.value)
    ...
    ... somewhere below ...
    ...
    file_option = Option(...)

I have also changed grouped all the calls to optionparser.py in
createQcdCmdLineParser(), which will return a parser, that will be called
afterwards.

Last, I using a pseudo "main" to separate the declarations from the actual
script imperative commands. Also this is good, if we want to import this script
and do some tests on it.