Refactor the way the options are handled by the opt parser.
After this change the OptionParser has the implementation for the syntax error
function, which is passed (by using the object reference) to the cmd_option
callbacks in the parsing/actions stage.

I have substituted the OptionParser::add method by a couple of methods
addConfiguration and addCommand, which are used from the program, and will act
as factory methods which will bild the appropriate option and store it in the
OptionParser::options list for later use.

I have also added a method, getOption, which will look for an Option with a
certain key at the OptionParser::options list.

Lasty I have introduced a first fix for avoiding the problem with bdb that adds
a extension ".db" to the created file, which made, all subsequent executions to
fail.