Fix listing problem in Mac OS
By default, in Mac Os, the db used by anydb does not support itemize, so the
list command didn't work.

With the fix, everytime one need to list the commands, if the databse selected
by anydbm to create the db file, is dumbdb or bsddb185 (as I had in my
machine), then I'll use list comprehensions to create the same output as
itemize.