Fix: qcd may migrate host databases when is not needed.
The idea of using a glob to get a set of valid (existent) qcd databases, was
not working correctly because qcd was matching the files with a rather
permissive regular expression.

The problem manifested itself, when being in machine B, with a shared $HOME for
two machines A and B, and a .qcddb.A database in this $HOME. In this case, if I
executed qcd, the script would rename .qcddb.A into .qcddb.B, which was an
incorrect behavior.