Add back flexibility when choosing the database name
qcd allows the user to define the file name of the database for qcd. When I
introduced the support for multiple machines with the same host, I forgot about
it, and forced it to be "\.qcddb.*". I have change the pattern now, so that it
becomes as flexible as before.